+++ pages/_document.tsx
@@ -79,7 +79,7 @@ export default class MyDocument extends Document {
           <link href="/favicon/site.webmanifest" rel="manifest" />
           <link color="#333" href="/favicon/safari-pinned-tab.svg" rel="mask-icon" />
           <meta content="#333" name="msapplication-TileColor" />
-          <meta content="#333" name="theme-color"></meta>
+          <meta content="#333" name="theme-color" />
         </Head>
         <body>
           <Main />
